Refactor to reduce the size of the utils package. This package seems to be doing a lot of different things, which goes against the Single Responsibility Principle. So I moved the functions related to importing and exporting blockchain data to a separate package, blockchainio. This to make the code easier to navigate and maintain for future devs, as it is clearer what each package is responsible for. It also makes it easier to test the code, as each package could be tested in isolation.
